Move enum to more appropriate header.
